+++ tools/assembler/asm_error.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <filesystem>
 #include <format>
 #include <stdexcept>
@@ -7,7 +8,7 @@
 
 class AsmError : public std::runtime_error {
   public:
-    explicit AsmError(int lineno, std::string_view message, const std::filesystem::path &origin = {})
+    explicit AsmError(size_t lineno, std::string_view message, const std::filesystem::path &origin = {})
         : std::runtime_error(origin.empty() ? std::format("{}, {}", lineno, message)
                                             : std::format("{}:{}, {}", origin.string(), lineno, message)) {}
 };
